Update: topDrivers report
