+++ src/components/BusinessIntelligence/ReportsSpentTIme/index.js
@@ -7,6 +7,7 @@ import { AnalyticsCalendar } from '../AnalyticsCalendar'
 import { Modal } from '../../Shared'
 import { ReportsDriverGroupFilter } from '../ReportsDriverGroupFilter'
 import { ReportsDriverFilter } from '../ReportsDriverFilter'
+import { convertHMS } from '../../../utils'
 import {
   ReportsBusinessSpendContainer,
   Title,
@@ -84,18 +85,6 @@ const ReportsSpentTimeUI = (props) => {
     document.body.appendChild(downloadLink)
     downloadLink.click()
     document.body.removeChild(downloadLink)
-  }
-
-  const convertHMS = (value) => {
-    const sec = parseInt(value, 10) // convert value to number if it's string
-    let hours = Math.floor(sec / 3600) // get hours
-    let minutes = Math.floor((sec - (hours * 3600)) / 60) // get minutes
-    let seconds = sec - (hours * 3600) - (minutes * 60) //  get seconds
-    // add 0 if value < 10; Example: 2 => 02
-    if (hours < 10) { hours = '0' + hours }
-    if (minutes < 10) { minutes = '0' + minutes }
-    if (seconds < 10) { seconds = '0' + seconds }
-    return hours + ':' + minutes + ':' + seconds // Return is HH : MM : SS
   }
 
   return (
